Move TypedServer from the imported example to lib/
